--- app/src/main/res/drawable/test.xml
@@ -1,21 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="100dp"
-    android:height="44dp"
-    android:viewportWidth="100"
-    android:viewportHeight="44">
-
-    <!-- 테두리 -->
-    <path
-        android:pathData="M2,2 L98,2 L98,42 L2,42 Z"
-        android:strokeWidth="3"
-        android:strokeColor="#FFDD00"
-        android:fillColor="@android:color/transparent"
-        android:fillType="nonZero" />
-
-    <!-- + 아이콘 -->
-    <path
-        android:pathData="M50,15 L50,29 M25,22 L75,22"
-        android:strokeColor="#FFDD00"
-        android:strokeWidth="4"
-        android:strokeLineCap="round" />
-</vector>
